mldsa fix off-by-one bug
the hint bitpacking implementation uses arithmetic instead of branching
to conditionally assign hint indices.
this creates an edgecase where the number of hints is equal to omega
causing the sum of hints for the first polynomial to be overridden.
the fix is the require the number of hints to be strictly less than
omega, which is not strictly compliant to ML-DSA but for now i consider
it preferable to introducing branching in the bitpacking implementation